fixed it so when the user who is logged in selects remove on either task or memo it will show a different verification screen. Next up will be tasks, instead of just removing it I will add a way that an employees name will show up

+++ src/Components/Item/ItemList.tsx
@@ -89,14 +89,18 @@ export const ItemListLayout: React.FC<{
   };
 
   const onRemove = (item: Item) => {
-
-    ItemService.delete(item.itemNumber)
-      .then((postResponse: any) => {
-        removeItem(item);
-      })
-      .catch((err: any) => {
-        alert("testing");
-      });
+    const removeItemConfirm = window.confirm("Are you sure you want to remove this item?")
+    if (!removeItemConfirm) {
+      return
+    } {
+      ItemService.delete(item.itemNumber)
+        .then((postResponse: any) => {
+          removeItem(item);
+        })
+        .catch((err: any) => {
+          alert("testing");
+        });
+    }
   };
 
   const loggedIn = () => {
